refactor: move config to config package
refactor: moved cluster logic to cluster.go
